+++ LED_control.h
@@ -4,9 +4,11 @@
  *
  *
 */
+#include "stdutils.h"
+
 extern void LED_setup(void);
 
-extern void LED_SET(unsigned int led_0 , unsigned int led_1, unsigned int led_2, unsigned int led_3);
+extern void LED_SET(uint8_t led_0 , uint8_t led_1, uint8_t led_2, uint8_t led_3);
 
 extern void disco_func(void );
